# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API Pedro",
        "description": "REST API -> TCC of Pedro Rodrigues Pereira",
        "contact": {
            "email": "jierlausson@hotmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/class-rooms": {
            "get": {
                "tags": [
                    "ClassRooms"
                ],
                "summary": "List all classrooms",
                "operationId": "766c98f5bcc5ee6ee92613506f431a1c",
                "responses": {
                    "200": {
                        "description": "List of classrooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClassRoom"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ClassRooms"
                ],
                "summary": "Create a new classroom",
                "operationId": "223f3e5bef66cc938217bbf04f69bc6c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sub_activity_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "unit_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "weight": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "max": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "class": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "class_time": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "day_month_year": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Classroom created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassRoom"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/class-rooms/{classRoom}": {
            "get": {
                "tags": [
                    "ClassRooms"
                ],
                "summary": "Retrieve a specific classroom",
                "operationId": "f6329f88f51bfb80d0eba962fe7d19da",
                "parameters": [
                    {
                        "name": "classRoom",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Classroom details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassRoom"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ClassRooms"
                ],
                "summary": "Update a classroom",
                "operationId": "01868ee1b5f8840c44d50502bca22fd2",
                "parameters": [
                    {
                        "name": "classRoom",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sub_activity_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "unit_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "weight": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "max": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "class": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "class_time": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "day_month_year": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Classroom updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassRoom"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ClassRooms"
                ],
                "summary": "Delete a classroom",
                "operationId": "364fa9507f58d2fdc2f5371a7fbf647b",
                "parameters": [
                    {
                        "name": "classRoom",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Classroom deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClassRoom": {
                "title": "ClassRoom",
                "description": "Modelo da sala de aula",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "description": "ID da sala de aula",
                        "type": "integer"
                    },
                    "sub_activity_id": {
                        "description": "ID da subatividade associada",
                        "type": "integer",
                        "nullable": true
                    },
                    "unit_id": {
                        "description": "ID da unidade associada",
                        "type": "integer",
                        "nullable": true
                    },
                    "description": {
                        "description": "Descrição da sala de aula",
                        "type": "string",
                        "nullable": true
                    },
                    "weight": {
                        "description": "Peso da sala de aula",
                        "type": "integer",
                        "nullable": true
                    },
                    "max": {
                        "description": "Valor máximo da sala",
                        "type": "integer",
                        "nullable": true
                    },
                    "class": {
                        "description": "Classe",
                        "type": "string",
                        "nullable": true
                    },
                    "class_time": {
                        "description": "Tempo de aula",
                        "type": "integer",
                        "nullable": true
                    },
                    "day_month_year": {
                        "description": "Data associada",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}